requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
colorama==0.4.6
decorator==5.1.1
Django==4.2.16
django-taggit==6.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
executing==2.1.0
factory-boy==3.3.1
Faker==33.0.0
importlib-resources==6.4.5
inflection==0.5.1
ipython==8.12.3
jedi==0.19.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
matplotlib-inline==0.1.7
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
pkgutil-resolve-name==1.3.10
prompt-toolkit==3.0.48
pure-eval==0.2.3
pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.1
six==1.16.0
sqlparse==0.5.2
stack-data==0.6.3
traitlets==5.14.3
typing-extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
uritemplate==4.1.1
wcwidth==0.2.13
zipp==3.20.2
